requirements.txt
torch>=2.0
torchvision
transformers>=4.30
numpy
Pillow
scikit-learn
matplotlib
seaborn
